apps/accounts/serializers.py
from rest_framework import serializers
from apps.common.constants import UserRole


class SendOTPSerializer(serializers.Serializer):
    email = serializers.EmailField()
    role = serializers.ChoiceField(choices=[r.value for r in UserRole])


class VerifyOTPSerializer(serializers.Serializer):
    email = serializers.EmailField()
    otp = serializers.CharField(max_length=6)