requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
django-environ==0.13.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dnspython==2.8.0
drf-yasg==1.21.15
gunicorn==25.1.0
inflection==0.5.1
mongoengine==0.29.1
packaging==26.0
pillow==12.1.1
PyJWT==2.11.0
pymongo==4.16.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.5
uritemplate==4.2.0
